refactor(instructions): replace getCompleteInstructions with getInstructions for improved modularity
