+++ src/utils/instructions/index.ts
@@ -1,0 +1,1 @@
+export { getInstructions } from './getInstructions.js';
