+++ src/utils/instructions/localFileUtils.ts
@@ -1,6 +1,6 @@
 import { promises as fs } from 'fs';
 import path from 'path';
-import type { ValidationResult } from '../types/validationResult.js';
+import type { ValidationResult } from '../../types/validationResult.js';
 
 /**
  * Validate if a file path has a markdown extension
